Move ExcisionSphere from DomainStructure to Domain
Coordinate maps will be added to ExcisionSphere so we don't want
cyclic dependencies
