+++ src/Domain/Creators/CylindricalBinaryCompactObject.cpp
@@ -22,12 +22,12 @@
 #include "Domain/Creators/BinaryCompactObjectHelpers.hpp"
 #include "Domain/Creators/ExpandOverBlocks.hpp"
 #include "Domain/DomainHelpers.hpp"
+#include "Domain/ExcisionSphere.hpp"
 #include "Domain/FunctionsOfTime/FixedSpeedCubic.hpp"
 #include "Domain/FunctionsOfTime/PiecewisePolynomial.hpp"
 #include "Domain/FunctionsOfTime/QuaternionFunctionOfTime.hpp"
 #include "Domain/Structure/Direction.hpp"
 #include "Domain/Structure/DirectionMap.hpp"
-#include "Domain/Structure/ExcisionSphere.hpp"
 #include "Domain/Structure/OrientationMap.hpp"
 #include "NumericalAlgorithms/RootFinding/QuadraticEquation.hpp"
 #include "Options/ParseError.hpp"
