+++ src/Domain/Structure/CMakeLists.txt
@@ -14,7 +14,6 @@ spectre_target_sources(
   Direction.cpp
   Element.cpp
   ElementId.cpp
-  ExcisionSphere.cpp
   Hypercube.cpp
   InitialElementIds.cpp
   Neighbors.cpp
@@ -38,7 +37,6 @@ spectre_target_headers(
   DirectionMap.hpp
   Element.hpp
   ElementId.hpp
-  ExcisionSphere.hpp
   Hypercube.hpp
   IndexToSliceAt.hpp
   InitialElementIds.hpp
