+++ src/Evolution/Systems/CurvedScalarWave/Worldtube/ElementActions/SendToWorldtube.hpp
@@ -15,8 +15,8 @@
 #include "DataStructures/Tensor/Slice.hpp"
 #include "DataStructures/Variables.hpp"
 #include "Domain/AreaElement.hpp"
+#include "Domain/ExcisionSphere.hpp"
 #include "Domain/Structure/Direction.hpp"
-#include "Domain/Structure/ExcisionSphere.hpp"
 #include "Domain/Structure/IndexToSliceAt.hpp"
 #include "Domain/Tags.hpp"
 #include "Domain/TagsTimeDependent.hpp"
