+++ src/Evolution/Systems/CurvedScalarWave/Worldtube/Tags.cpp
@@ -12,8 +12,8 @@
 #include "DataStructures/Tensor/Slice.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Variables.hpp"
+#include "Domain/ExcisionSphere.hpp"
 #include "Domain/Structure/Element.hpp"
-#include "Domain/Structure/ExcisionSphere.hpp"
 #include "Domain/Structure/IndexToSliceAt.hpp"
 #include "Evolution/Systems/CurvedScalarWave/Worldtube/PunctureField.hpp"
 #include "NumericalAlgorithms/Spectral/Mesh.hpp"
