+++ tests/Unit/Domain/CMakeLists.txt
@@ -20,6 +20,7 @@ set(LIBRARY_SOURCES
   Test_ElementDistribution.cpp
   Test_ElementMap.cpp
   Test_ElementToBlockLogicalMap.cpp
+  Test_ExcisionSphere.cpp
   Test_FaceNormal.cpp
   Test_FlatLogicalMetric.cpp
   Test_InterfaceHelpers.cpp
