+++ tests/Unit/Domain/Structure/CMakeLists.txt
@@ -11,7 +11,6 @@ set(LIBRARY_SOURCES
   Test_Direction.cpp
   Test_Element.cpp
   Test_ElementId.cpp
-  Test_ExcisionSphere.cpp
   Test_Hypercube.cpp
   Test_IndexToSliceAt.cpp
   Test_InitialElementIds.cpp
