+++ tests/Unit/Domain/Test_AreaElement.cpp
@@ -21,9 +21,9 @@
 #include "Domain/Creators/Sphere.hpp"
 #include "Domain/ElementMap.hpp"
 #include "Domain/ElementToBlockLogicalMap.hpp"
+#include "Domain/ExcisionSphere.hpp"
 #include "Domain/FaceNormal.hpp"
 #include "Domain/InterfaceLogicalCoordinates.hpp"
-#include "Domain/Structure/ExcisionSphere.hpp"
 #include "Domain/Structure/InitialElementIds.hpp"
 #include "Domain/Structure/OrientationMap.hpp"
 #include "Helpers/DataStructures/MakeWithRandomValues.hpp"
